Fixed "adding to path" for windows after rewrite

+++ volpe/__main__.py
@@ -22,33 +22,21 @@ def install():
     path_to_volpe_root = os.path.dirname(path_to_volpe)
     path_to_volpe_bin = os.path.join(path_to_volpe_root, "bin")
 
-    path = os.environ["PATH"]
-
-    if path_to_volpe_bin in path:
-        print("Volpe is already on PATH.")
-        return
+    batch_file_path = os.path.join(path_to_volpe_bin, "volpe.bat")
+    bash_file_path = os.path.join(path_to_volpe_bin, "volpe")
 
     if not os.path.isdir(path_to_volpe_bin):
         os.makedirs(path_to_volpe_bin)
 
-    if platform == "win32":
+    if platform == "win32" and not os.path.isfile(batch_file_path):
         # Update the batch file.
-        print("Updating batch file with volpe directory.")
-        batch_file_path = os.path.join(path_to_volpe_bin, "volpe.bat")
+        print("Creating batch file with volpe directory.")
         with open(batch_file_path, "w") as batch_file:
             batch_file.writelines(["@echo off\n", f"python {path_to_volpe} %*"])
-  
-        # Add Volpe path to user path.
-        print(f"Adding {path_to_volpe_bin} to user PATH.")
-        path_including_volpe = path + os.pathsep + path_to_volpe_root
-        os.system(f'SETX Path "{path_including_volpe}"')
 
-        print("Please restart this console for changes to take effect.")
-
-    elif platform == "linux":
+    elif platform == "linux" and not os.path.isfile(bash_file_path):
         # Update the bash file.
-        print("Updating bash file with volpe directory.")
-        bash_file_path = os.path.join(path_to_volpe_bin, "volpe")
+        print("Creating bash file with volpe directory.")
         with open(bash_file_path, "w") as bash_file:
             bash_file.writelines(["#!/bin/bash\n", f"python {path_to_volpe} $@"])
 
@@ -58,13 +46,44 @@ def install():
         st = os.stat(bash_file_path)
         os.chmod(bash_file_path, st.st_mode | stat.S_IXUSR | stat.S_IXGRP | stat.S_IXOTH)
 
+    if path_to_volpe_bin in os.environ["PATH"]:
+        print("Volpe is already on PATH.")
+        return
+
+    if platform == "win32":
+        # Make sure it is not on PATH before trying to add it
+        import winreg
+        reg_key = winreg.OpenKey(winreg.HKEY_CURRENT_USER, "Environment")
+        path = winreg.QueryValueEx(reg_key, "Path")[0]
+
+        if path_to_volpe_bin not in path:
+            with open("PATH_BACKUP.txt", "w") as f:
+                print("Writing current PATH to PATH_BACKUP.txt in case it gets accidentally truncated.")
+                f.write(path)
+            # Add Volpe path to user path.
+            print(f"Adding {path_to_volpe_bin} to user PATH.")
+            path_including_volpe = path + os.pathsep + path_to_volpe_bin
+            os.system(f'SETX Path "{path_including_volpe}"')
+
+        print("Please restart this console for changes to take effect.")
+
+    elif platform == "linux":
         # Add Volpe path to profile.
         home = os.path.expanduser("~")
         profile = os.path.join(home, ".profile")
         command = f"export PATH=\"$PATH:{path_to_volpe_bin}\""
-        print(f"Adding {command} to {profile}")
-        with open(profile, "a") as profile_file:
-            profile_file.write(f"\n{command}\n")
+
+        with open(profile, "r") as profile_file:
+            contents = profile_file.read()
+            already_there = command in contents
+            needs_newline = contents[-1] == "\n"
+
+        if not already_there:
+            print(f"Adding {command} to {profile}")
+            with open(profile, "a") as profile_file:
+                if needs_newline:
+                    profile_file.write("\n")
+                profile_file.write(f"{command}\n")
 
         print("Please reboot for changes to take effect.")
